alarm created, also some formatting

+++ spring/fmwk/src/main/java/Clock.java
@@ -3,5 +3,5 @@
  */
 public interface Clock {
 
-    Integer whatTimeIsIt();
+    Long whatTimeIsIt();
 }
